+++ shared/types/src/elastic/editorial-content.ts
@@ -1,0 +1,7 @@
+import { EditorialContentDoc } from "../hasura";
+import { KeysToCamelCase } from "../utility";
+import { DocumentElasticWithSource } from "./common";
+
+export type EditorialContentElasticDocument = DocumentElasticWithSource<
+  KeysToCamelCase<EditorialContentDoc>
+>;
