+++ shared/types/src/elastic/index.ts
@@ -1,4 +1,6 @@
 export * from "./agreements";
 export * from "./contributions";
-export * from "./global";
+export * from "./common";
 export * from "./prequalified";
+export * from "./tools";
+export * from "./editorial-content";
