+++ shared/types/src/elastic/prequalified.ts
@@ -1,0 +1,7 @@
+import { Prequalified } from "../hasura";
+import { DocumentElasticWithSource } from "./common";
+
+export type PrequalifiedElasticDocument = Omit<
+  DocumentElasticWithSource<Prequalified>,
+  "slug"
+>;
