--- shared/types/src/elasticsearch/agreement.ts
@@ -1,14 +1,0 @@
-export type Agreement = {
-  id: string;
-  num: number;
-  shortTitle: string;
-  slug?: string;
-  title: string;
-  url?: string;
-  highlight?: {
-    title: string;
-    content: string;
-    searchInfo?: string;
-  };
-  contributions: boolean;
-};
