--- shared/types/src/elasticsearch/index.ts
@@ -1,5 +1,0 @@
-export * from "./common";
-export * from "./editorial-content";
-export * from "./tool";
-export * from "./agreement";
-export * from "./contribution";
