--- shared/types/src/export/prequalified.ts
@@ -1,11 +1,0 @@
-import { DocumentElasticWithSource } from "./global";
-
-type Prequalified = {
-  source: "prequalified";
-  variants: string[];
-};
-
-export type PrequalifiedElasticDocument = Omit<
-  DocumentElasticWithSource<Prequalified>,
-  "slug"
->;
