+++ shared/types/src/hasura/highlights.ts
@@ -1,0 +1,3 @@
+import { HasuraDocument } from "./common";
+
+export type Highlight = HasuraDocument<any, "highlights">;
