+++ shared/types/src/hasura/index.ts
@@ -1,0 +1,14 @@
+export * from "./agreement";
+export * from "./alert";
+export * from "./code-du-travail";
+export * from "./common";
+export * from "./contributions";
+export * from "./dila-resolver";
+export * from "./editorial-content";
+export * from "./export-es";
+export * from "./fiche-sp";
+export * from "./fiche-travail";
+export * from "./highlights";
+export * from "./modeles-de-courrier";
+export * from "./prequalified";
+export * from "./themes";
