+++ shared/types/src/hasura/prequalified.ts
@@ -1,0 +1,7 @@
+import { HasuraDocument } from "./common";
+
+export type PrequalifiedDoc = {
+  variants: string[];
+};
+
+export type Prequalified = HasuraDocument<PrequalifiedDoc, "prequalified">;
