+++ shared/types/src/hasura/themes.ts
@@ -1,0 +1,9 @@
+import { HasuraDocument } from "./common";
+
+export type Theme = HasuraDocument<ThemeDoc, "themes">;
+
+export type ThemeDoc = {
+  icon?: string;
+  shortTitle?: string;
+  description?: string;
+};
