+++ shared/types/src/utility.ts
